:recycle: refactor: Re-structuring the files for a future API Implementation and better architecture
