+++ src/Tests/StatementPrinterTests.TestStatementExampleLegacy.approved.txt
