+++ src/Tests/StatementPrinterTests.TestTextStatementExample.approved.txt
